+++ test/e2e/framework/nodes_util.go
@@ -16,21 +16,6 @@ limitations under the License.
 
 package framework
 
-import (
-	"sync"
-	"time"
-
-	"github.com/onsi/ginkgo/v2"
-
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/wait"
-	clientset "k8s.io/client-go/kubernetes"
-
-	// TODO: Remove the following imports (ref: https://github.com/kubernetes/kubernetes/issues/81245)
-	e2enode "k8s.io/kubernetes/test/e2e/framework/node"
-	e2essh "k8s.io/kubernetes/test/e2e/framework/ssh"
-)
-
 // AppendContainerCommandGroupIfNeeded returns container command group parameter if necessary.
 func AppendContainerCommandGroupIfNeeded(args []string) []string {
 	if TestContext.CloudConfig.Region != "" {
@@ -38,66 +23,4 @@ func AppendContainerCommandGroupIfNeeded(args []string) []string {
 		return append([]string{"beta"}, args...)
 	}
 	return args
-}
-
-// NodeKiller is a utility to simulate node failures.
-type NodeKiller struct {
-	config   NodeKillerConfig
-	client   clientset.Interface
-	provider string
-}
-
-// NewNodeKiller creates new NodeKiller.
-func NewNodeKiller(config NodeKillerConfig, client clientset.Interface, provider string) *NodeKiller {
-	config.NodeKillerStopCh = make(chan struct{})
-	return &NodeKiller{config, client, provider}
-}
-
-// Run starts NodeKiller until stopCh is closed.
-func (k *NodeKiller) Run(stopCh <-chan struct{}) {
-	// wait.JitterUntil starts work immediately, so wait first.
-	time.Sleep(wait.Jitter(k.config.Interval, k.config.JitterFactor))
-	wait.JitterUntil(func() {
-		nodes := k.pickNodes()
-		k.kill(nodes)
-	}, k.config.Interval, k.config.JitterFactor, true, stopCh)
-}
-
-func (k *NodeKiller) pickNodes() []v1.Node {
-	nodes, err := e2enode.GetReadySchedulableNodes(k.client)
-	ExpectNoError(err)
-	numNodes := int(k.config.FailureRatio * float64(len(nodes.Items)))
-
-	nodes, err = e2enode.GetBoundedReadySchedulableNodes(k.client, numNodes)
-	ExpectNoError(err)
-	return nodes.Items
-}
-
-func (k *NodeKiller) kill(nodes []v1.Node) {
-	wg := sync.WaitGroup{}
-	wg.Add(len(nodes))
-	for _, node := range nodes {
-		node := node
-		go func() {
-			defer ginkgo.GinkgoRecover()
-			defer wg.Done()
-
-			Logf("Stopping docker and kubelet on %q to simulate failure", node.Name)
-			err := e2essh.IssueSSHCommand("sudo systemctl stop docker kubelet", k.provider, &node)
-			if err != nil {
-				Logf("ERROR while stopping node %q: %v", node.Name, err)
-				return
-			}
-
-			time.Sleep(k.config.SimulatedDowntime)
-
-			Logf("Rebooting %q to repair the node", node.Name)
-			err = e2essh.IssueSSHCommand("sudo reboot", k.provider, &node)
-			if err != nil {
-				Logf("ERROR while rebooting node %q: %v", node.Name, err)
-				return
-			}
-		}()
-	}
-	wg.Wait()
 }
